backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
requests
openai
pyjwt
google-cloud-storage
pinecone-client
sqlite3
# For multimodal providers
google-generativeai
boto3
sentence-transformers
Pillow
torch
